Preserve decimals on amount (f32 instead of u32)

+++ packages/anoma-lib/lib/src/ibc_transfer.rs
@@ -33,7 +33,7 @@ impl IbcTransfer {
         sender: String,
         receiver: String,
         token: String,
-        amount: u32,
+        amount: f32,
         epoch: u32,
         fee_amount: u32,
         gas_limit: u32,
